model_serving/requirements.txt
uvicorn
fastapi
google-cloud-bigquery
pandas
requests
python-dotenv
pyarrow
konlpy
pyspark==3.4.1
scikit-learn
